+++ src/components/modelos/CrearModeloModal.tsx
@@ -1,60 +1,30 @@
 'use client';
+import FormCrearModelo from '@/components/modelos/FormCrearModelo';
+import { trpc } from '@/lib/trpc';
+import { ModelosSimilarity } from '@/server/types/modelos';
+import clsx from 'clsx';
 import { toast } from 'sonner';
 import { Button } from '../ui/button';
 import { Dialog, DialogContent, DialogTrigger } from '../ui/dialog';
-import { Input } from '../ui/input';
-import { useCrearModeloModal } from './CrearModelo';
-import { trpc } from '@/lib/trpc';
 import Loader from '../ui/loader';
-import CirclePlus from '../icons/CirclePlus';
-import { useMemo, useRef, useState } from 'react';
-import { RouterOutputs } from '@/server';
-import { Label } from '../ui/label';
-import {
-  Select,
-  SelectContent,
-  SelectItem,
-  SelectTrigger,
-  SelectValue,
-} from '../ui/select';
-import { Badge } from '../ui/badge';
-import { getTextColorByBg } from '@/lib/utils';
-import CircleXIcon from '../icons/CircleX';
-import ComboBox from '../ui/ComboBox';
-import EtiquetasFillIcon from '../icons/EtiquetasFillIcon';
-import { TrashIcon } from 'lucide-react';
-import { ModelosSimilarity } from '@/server/types/modelos';
-import ModeloFillIcon from '../icons/ModeloFillIcon';
-import { useRouter } from 'next/navigation';
+import { useCrearModeloModal } from './CrearModelo';
+import { useRef, useState } from 'react';
+import ModelosSimilares from '@/components/modelos/ModelosSimilares';
 
 const CrearModeloModal = ({ open }: { open: boolean }) => {
   const modalModelo = useCrearModeloModal();
   const utils = trpc.useUtils();
-  const router = useRouter();
   const createModelo = trpc.modelo.createManual.useMutation();
+
   const inputRef = useRef<HTMLInputElement>(null);
   const [video, setVideo] = useState<File | null>(null);
   const [fotoUrl, setFotoUrl] = useState<string | null>(null);
+
   const [similarity, setSimilarity] = useState<boolean>(false);
   const [similarityView, setSimilarityView] = useState<boolean>(false);
   const [similarityModelos, setSimilarityModelos] = useState<ModelosSimilarity>(
     []
   );
-  const [openSelect, setOpenSelect] = useState(false);
-  const [addEtiquetaOpen, setAddEtiquetaOpen] = useState(false);
-  const [comboBoxGrupoOpen, setComboBoxGrupoOpen] = useState(false);
-  const [comboBoxEtiquetaOpen, setComboBoxEtiquetaOpen] = useState(false);
-  const { data: grupoEtiquetas } = trpc.grupoEtiqueta.getAll.useQuery();
-  const [grupoEtiquetaSelected, setGrupoEtiquetaSelected] =
-    useState<string>('');
-  const [etiquetaSelected, setEtiquetaSelected] = useState<string>('');
-  const { data: etiquetasGrupo } =
-    grupoEtiquetaSelected === ''
-      ? trpc.etiqueta.getAll.useQuery()
-      : trpc.etiqueta.getByGrupoEtiqueta.useQuery(grupoEtiquetaSelected);
-  const currentGrupo = useMemo(() => {
-    return grupoEtiquetas?.find((g) => g.id === grupoEtiquetaSelected);
-  }, [grupoEtiquetas, grupoEtiquetaSelected]);
 
   async function handleSave() {
     if (modalModelo.modelo.nombreCompleto === '') {
@@ -81,24 +51,12 @@ const CrearModeloModal = ({ open }: { open: boolean }) => {
 
     if (Array.isArray(res)) {
       setSimilarity(true);
-      console.log('similarity', similarity);
-      res.forEach((r) => {
-        // Cambiar la fecha de nacimiento de Date a string
-        const fechanacimiento = r.modelo.fechaNacimiento
-          ? new Date(r.modelo.fechaNacimiento)
-          : null;
-      });
       setSimilarityModelos(
         res.map((r) => ({
           similarityTelefono: r.similarityTelefono,
           similarityNombre: r.similarityNombre,
           modelo: {
             ...r.modelo,
-            fechaNacimiento: r.modelo.fechaNacimiento
-              ? new Date(r.modelo.fechaNacimiento)
-              : null,
-            created_at: new Date(r.modelo.created_at),
-            updated_at: new Date(r.modelo.updated_at),
           },
         }))
       );
@@ -164,68 +122,7 @@ const CrearModeloModal = ({ open }: { open: boolean }) => {
     });
     setVideo(null);
     setFotoUrl(null);
-    setGrupoEtiquetaSelected('');
-    setEtiquetaSelected('');
-    setAddEtiquetaOpen(false);
     setSimilarity(false);
-    // inputRef.current!.value = '';
-  }
-  async function handleDeleteEtiqueta(
-    etiqueta: NonNullable<RouterOutputs['etiqueta']['getById']>
-  ) {
-    useCrearModeloModal.setState({
-      modelo: {
-        ...modalModelo.modelo,
-        etiquetas: modalModelo.modelo.etiquetas.filter(
-          (e) => e.id !== etiqueta.id
-        ),
-      },
-    });
-  }
-  async function handleDeleteNombre(index: number) {
-    const newApodos = modalModelo.modelo.apodos;
-    newApodos.splice(index, 1);
-    useCrearModeloModal.setState({
-      modelo: {
-        ...modalModelo.modelo,
-        apodos: newApodos,
-      },
-    });
-  }
-  async function handleAddEtiqueta() {
-    if (etiquetaSelected === '') return;
-    const etiqueta = etiquetasGrupo?.find((e) => e.id === etiquetaSelected);
-    if (
-      useCrearModeloModal
-        .getState()
-        .modelo.etiquetas.find((e) => e.id === etiqueta?.id)
-    ) {
-      toast.error('La etiqueta ya fue agregada');
-      return;
-    }
-    if (!etiqueta) return;
-
-    // Verificar grupo exclusivo o no
-    if (
-      etiqueta.grupo.esExclusivo &&
-      useCrearModeloModal
-        .getState()
-        .modelo.etiquetas.find(
-          (e) => e.grupo.id === etiqueta?.grupo.id && e.id !== etiqueta?.id
-        )
-    ) {
-      toast.error('No puedes agregar dos etiquetas exclusivas del mismo grupo');
-      return;
-    }
-    useCrearModeloModal.setState({
-      modelo: {
-        ...modalModelo.modelo,
-        etiquetas: [...modalModelo.modelo.etiquetas, etiqueta],
-      },
-    });
-    setEtiquetaSelected('');
-    setGrupoEtiquetaSelected('');
-    setAddEtiquetaOpen(false);
   }
 
   return (
@@ -239,348 +136,55 @@ const CrearModeloModal = ({ open }: { open: boolean }) => {
         <DialogTrigger></DialogTrigger>
         <DialogContent onCloseAutoFocus={handleCancel} className=''>
           <div className='flex flex-col gap-y-0.5'>
-            <p className='text-base font-semibold'>
+            <p className='text-xl font-semibold'>
               Crear participante manualmente
             </p>
             {!similarityView ? (
               <div className='mt-1 flex max-h-[400px] flex-col gap-y-1 overflow-y-auto px-2'>
-                <Label className='text-sm'>
-                  Nombre completo: (obligatorio)
-                </Label>
-                <Input
-                  type='text'
-                  placeholder='Nombre Completo'
-                  className=''
-                  value={modalModelo.modelo.nombreCompleto}
-                  onChange={(e) =>
-                    useCrearModeloModal.setState({
-                      modelo: {
-                        ...modalModelo.modelo,
-                        nombreCompleto: e.target.value,
-                      },
-                    })
-                  }
-                  required
+                <FormCrearModelo
+                  video={video}
+                  setVideo={setVideo}
+                  fotoUrl={fotoUrl}
+                  setFotoUrl={setFotoUrl}
+                  inputRef={inputRef}
                 />
-                <Label className='pt-2 text-sm'>Teléfono: (obligatorio)</Label>
-                <Input
-                  type='text'
-                  placeholder='Teléfono'
-                  className=''
-                  value={modalModelo.modelo.telefono}
-                  onChange={(e) =>
-                    useCrearModeloModal.setState({
-                      modelo: {
-                        ...modalModelo.modelo,
-                        telefono: e.target.value,
-                      },
-                    })
-                  }
-                  required
-                />
-                <Label className='pt-2 text-sm'>DNI:</Label>
-                <Input
-                  type='text'
-                  placeholder='DNI'
-                  className=''
-                  value={modalModelo.modelo.dni}
-                  onChange={(e) =>
-                    useCrearModeloModal.setState({
-                      modelo: { ...modalModelo.modelo, dni: e.target.value },
-                    })
-                  }
-                />
-                <Label className='pt-2 text-sm'>Fecha de nacimiento:</Label>
-                <Input
-                  type='date'
-                  placeholder='Fecha de nacimiento'
-                  className='py-4'
-                  value={
-                    modalModelo.modelo.fechaNacimiento
-                      ?.toISOString()
-                      .split('T')[0]
-                  }
-                  onChange={(e) =>
-                    useCrearModeloModal.setState({
-                      modelo: {
-                        ...modalModelo.modelo,
-                        fechaNacimiento: new Date(e.currentTarget.value),
-                      },
-                    })
-                  }
-                />
-                <div>
-                  <Label htmlFor='genero'>Genero</Label>
-                  <Select
-                    open={openSelect}
-                    onOpenChange={setOpenSelect}
-                    onValueChange={(value) => {
-                      useCrearModeloModal.setState({
-                        modelo: {
-                          ...modalModelo.modelo,
-                          genero: value as string,
-                        },
-                      });
-                    }}
-                    defaultValue={modalModelo.modelo.genero ?? 'N/A'}
-                  >
-                    <SelectTrigger className=''>
-                      <SelectValue placeholder='Genero' />
-                    </SelectTrigger>
-                    <SelectContent>
-                      <SelectItem value='Femenino'>Femenino</SelectItem>
-                      <SelectItem value='Masculino'>Masculino</SelectItem>
-                      <SelectItem value='Otro'>Otro</SelectItem>
-                      <SelectItem value='N/A'>N/A</SelectItem>
-                    </SelectContent>
-                  </Select>
-                </div>
-                <Label className='pt-2 text-sm'>Mail:</Label>
-                <Input
-                  type='text'
-                  placeholder='Mail'
-                  className=''
-                  value={modalModelo.modelo.mail}
-                  onChange={(e) =>
-                    useCrearModeloModal.setState({
-                      modelo: { ...modalModelo.modelo, mail: e.target.value },
-                    })
-                  }
-                />
-                <Label className='pt-2 text-sm'>Instagram:</Label>
-                <div className='flex items-center justify-center gap-x-2'>
-                  <p className='text-xs'>instagram.com/</p>
-                  <Input
-                    type='text'
-                    placeholder='Instagram'
-                    className=''
-                    value={modalModelo.modelo.instagram}
-                    onChange={(e) =>
-                      useCrearModeloModal.setState({
-                        modelo: {
-                          ...modalModelo.modelo,
-                          instagram: e.target.value,
-                        },
-                      })
-                    }
-                  />
-                </div>
-                <Label className='pt-2 text-sm'>Foto:</Label>
-                <div className='flex gap-x-2'>
-                  <label className='flex aspect-square h-8 w-8 items-center justify-center rounded-full border-2 bg-black text-white hover:cursor-pointer'>
-                    <CirclePlus className='h-6 w-6 md:h-8 md:w-8' />
-                    <input
-                      type='file'
-                      name='imagen'
-                      className='hidden'
-                      accept='image/jpeg,image/png,image/webp'
-                      ref={inputRef}
-                      onChange={(e) => {
-                        const file = e.target.files?.[0];
-                        setVideo(file ?? null);
-                        setFotoUrl(!file ? null : URL.createObjectURL(file));
-                      }}
-                    />
-                  </label>
-                  {video && (
-                    <span className='mt-1 max-w-full truncate text-sm'>
-                      {video.name}
-                    </span>
-                  )}
-                </div>
-                <Label className='pt-2 text-sm'>Etiquetas:</Label>
-                <div className='flex flex-wrap items-center gap-2'>
-                  {modalModelo.modelo.etiquetas?.map((etiqueta) => {
-                    if (!etiqueta) return;
-                    return (
-                      <Badge
-                        className='group whitespace-nowrap transition-transform duration-200 ease-in-out hover:shadow-md'
-                        style={{
-                          backgroundColor: etiqueta?.grupo.color,
-                          color: getTextColorByBg(etiqueta.grupo.color),
-                        }}
-                        key={etiqueta.id}
-                      >
-                        {etiqueta.nombre}
-
-                        <CircleXIcon
-                          onClick={() => handleDeleteEtiqueta(etiqueta)}
-                          className='invisible w-0 cursor-pointer group-hover:visible group-hover:ml-1 group-hover:w-4'
-                          width={16}
-                          height={16}
-                        />
-                      </Badge>
-                    );
-                  })}
-                  {addEtiquetaOpen ? (
-                    <CircleXIcon
-                      onClick={() => setAddEtiquetaOpen(false)}
-                      className='h-5 w-5 cursor-pointer'
-                    />
-                  ) : (
-                    <CirclePlus
-                      className='h-5 w-5 cursor-pointer'
-                      onClick={() => setAddEtiquetaOpen(true)}
-                    />
-                  )}
-                </div>
-                {addEtiquetaOpen && (
-                  <div className='flex flex-wrap gap-x-2 gap-y-1'>
-                    <ComboBox
-                      open={comboBoxGrupoOpen}
-                      setOpen={setComboBoxGrupoOpen}
-                      value='nombre'
-                      id={'id'}
-                      data={grupoEtiquetas ?? []}
-                      onSelect={(selectedItem) => {
-                        if (selectedItem === grupoEtiquetaSelected) {
-                          setGrupoEtiquetaSelected('');
-                          setComboBoxGrupoOpen(false);
-                        } else {
-                          setGrupoEtiquetaSelected(selectedItem);
-                          setComboBoxGrupoOpen(false);
-                        }
-                      }}
-                      wFullMobile
-                      selectedIf={
-                        grupoEtiquetaSelected ? grupoEtiquetaSelected : ''
-                      }
-                      triggerChildren={
-                        <>
-                          <span className='max-w-[calc(100%-30px)] truncate'>
-                            {grupoEtiquetaSelected
-                              ? currentGrupo?.nombre
-                              : 'Buscar grupo...'}
-                          </span>
-                          <EtiquetasFillIcon className='h-5 w-5' />
-                        </>
-                      }
-                    />
-                    <ComboBox
-                      data={etiquetasGrupo ?? []}
-                      id={'id'}
-                      value='nombre'
-                      wFullMobile
-                      open={comboBoxEtiquetaOpen}
-                      setOpen={setComboBoxEtiquetaOpen}
-                      onSelect={(selectedItem) => {
-                        if (selectedItem === etiquetaSelected) {
-                          setEtiquetaSelected('');
-                          setComboBoxEtiquetaOpen(false);
-                        } else {
-                          setEtiquetaSelected(selectedItem);
-                          setComboBoxEtiquetaOpen(false);
-                        }
-                      }}
-                      selectedIf={etiquetaSelected}
-                      triggerChildren={
-                        <>
-                          <span className='truncate'>
-                            {etiquetaSelected !== ''
-                              ? etiquetasGrupo?.find(
-                                  (etiqueta) => etiqueta.id === etiquetaSelected
-                                )?.nombre ?? 'Buscar etiqueta...'
-                              : 'Buscar etiqueta...'}
-                          </span>
-                          <EtiquetasFillIcon className='h-5 w-5' />
-                        </>
-                      }
-                    />
-                    <Button onClick={handleAddEtiqueta}>Agregar</Button>
-                  </div>
-                )}
-                <Label className='pt-2 text-sm'>Nombres alternativos:</Label>
-                <div className='flex w-full flex-col items-start gap-y-1'>
-                  {modalModelo.modelo.apodos?.map((apodo, index) => {
-                    return (
-                      <div
-                        key={index}
-                        className='flex w-full items-center justify-between gap-x-3'
-                      >
-                        <Input
-                          type='text'
-                          placeholder='Nombre'
-                          value={apodo}
-                          key={index}
-                          onChange={(e) => {
-                            const newApodos = modalModelo.modelo.apodos;
-                            newApodos[index] = e.target.value;
-                            useCrearModeloModal.setState({
-                              modelo: {
-                                ...modalModelo.modelo,
-                                apodos: newApodos,
-                              },
-                            });
-                          }}
-                        />
-                        <TrashIcon
-                          onClick={() => handleDeleteNombre(index)}
-                          className='h-6 w-6 cursor-pointer'
-                          width={16}
-                          height={16}
-                        />
-                      </div>
-                    );
-                  })}
-                  <CirclePlus
-                    className='h-5 w-5 cursor-pointer'
-                    onClick={() => {
-                      useCrearModeloModal.setState({
-                        modelo: {
-                          ...modalModelo.modelo,
-                          apodos: [...modalModelo.modelo.apodos, ''],
-                        },
-                      });
-                    }}
-                  />
-                </div>
               </div>
             ) : (
               <div className='flex max-h-[400px] flex-col gap-y-2 overflow-y-auto'>
-                {similarityModelos.map((modelo) => (
-                  <div
-                    key={modelo.modelo.id}
-                    className='flex items-center justify-between'
-                  >
-                    <p>Nombre: {modelo.modelo.nombreCompleto}</p>
-                    <p>Teléfono: {modelo.modelo.nombreCompleto}</p>
-                    <ModeloFillIcon
-                      className='h-6 w-6 hover:text-gray-400'
-                      onClick={() => {
-                        router.push(`/modelo/${modelo.modelo.id}`);
-                      }}
-                    />
-                  </div>
-                ))}
+                <ModelosSimilares similarityModelos={similarityModelos} />
                 <Button onClick={() => setSimilarityView(false)}>Volver</Button>
               </div>
             )}
             <div
-              className={`flex ${similarity ? 'justify-between' : 'justify-end'} gap-x-2 pt-2`}
+              className={clsx(
+                `flex gap-x-2 pt-2`,
+                similarity ? 'justify-between' : 'justify-end'
+              )}
             >
               {similarity && (
                 <div>
-                  <span className='text-xs'>
-                    Hay {similarityModelos.length}{' '}
-                    {similarityModelos.length === 1
-                      ? 'modelo similar.'
-                      : 'modelos similares.'}{' '}
+                  <span className='align-middle text-xs'>
+                    Hay{' '}
+                    <span
+                      className='cursor-pointer font-semibold underline'
+                      onClick={() => setSimilarityView(true)}
+                    >
+                      {similarityModelos.length}{' '}
+                      {similarityModelos.length === 1
+                        ? 'modelo similar.'
+                        : 'modelos similares.'}
+                    </span>{' '}
                     ¿Quieres agregar a este participante?
                   </span>
-                  <Button
-                    className='text-xs'
-                    onClick={() => setSimilarityView(true)}
-                  >
-                    Ver los participantes similares
-                  </Button>
                 </div>
               )}
               <Button
                 onClick={handleSave}
                 className='flex justify-center gap-x-2'
+                disabled={createModelo.isLoading}
               >
                 {createModelo.isLoading ?? <Loader className='h-5 w-5' />}
-                <p>{similarity ? 'Agregar' : 'Guardar'}</p>
+                <p>{similarity ? 'Agregar igualmente' : 'Guardar'}</p>
               </Button>
             </div>
           </div>
