+++ src/server/types/modelos.ts
@@ -3,5 +3,5 @@ import { Perfil } from '@prisma/client';
 export type ModelosSimilarity = {
   similarityTelefono: number;
   similarityNombre: number;
-  modelo: Perfil;
+  modelo: Pick<Perfil, 'nombreCompleto' | 'id'>;
 }[];
